Implement optional cell height config
Now Source maintains a `cellHeightCache` which is set when the cell it
receives a callback from the cell that it has determined its height
via the completion handler in the Configurable protocol. This
cell-height cach has to be used by the UITableViewDelegate in
`tableView(_:, heightForRowAt:)`, i.e. in the respective view
controllers you use Source.