Implement Custom Section Headers
We now have headerModel and footerModel that are nil by default but can
be configured similar like cells using a custom model and view
implementation.
The only difference is that showing them is controlled via the
UITableViewDelegate so you have to implement this functionality inside
your app. But due to the new protocol SupplementaryViewModel and
ConfigurableSupplementaryView this is very easy. For an example you can
have a look at MultiSectionViewController from the example in Source.