Fix separator insets
From an old article I leaned to set the directionalLayoutMatrgins
for hiding the separator insets which is not correct anymore since
it pushes the complete cell content into a direction. Now by
removing this from the adjustment function the
directionalLayoutMargins can actually be used to control the cells
insets which is as intened.